requirements.txt
amqp==2.1.4
anyjson==0.3.3
appdirs==1.4.3
billiard==3.5.0.2
celery==4.0.2
configparser==3.5.0
Django==1.11.2
django-cors-headers==2.0.2
django-guardian==1.4.8
djangorestframework==3.6.2
kombu==4.0.2
olefile==0.44
packaging==16.8
Pillow==4.1.1
psycopg2==2.6.2
pyparsing==2.2.0
python-memcached==1.58
pytz==2017.2
redis==2.10.5
six==1.10.0
SQLAlchemy==1.1.9
vine==1.1.3
